Deprecate get_file_parser from modelicalang package
