+++ modelicalang/__init__.py
@@ -1,6 +1,5 @@
 __all__ = (
     "ModelicaVersion",
-    "get_file_parser",
     "get_syntax_type",
     "ParsingExpressionLike",
     "returns_parsing_expression",
@@ -13,7 +12,6 @@ import enum
 from functools import lru_cache
 from typing import TYPE_CHECKING, Dict, Optional, Type, Union
 
-from arpeggio import EOF, ParserPython
 from typing_extensions import TypeAlias
 
 if TYPE_CHECKING:
@@ -38,17 +36,6 @@ _AnySyntaxType = Union[Type[v3_4.Syntax], Type[v3_5.Syntax]]
 class ModelicaVersion(enum.Enum):
     v3_4 = (3, 4)
     v3_5 = latest = (3, 5)
-
-
-def get_file_parser(version: Optional[ModelicaVersion] = None) -> ParserPython:
-    syntax_type = get_syntax_type(version)
-
-    @returns_parsing_expression
-    def file() -> ParsingExpressionLike:
-        return syntax_type.stored_definition, EOF
-
-    with syntax_type:
-        return ParserPython(file, syntax_type.COMMENT)
 
 
 @lru_cache(1)
